split the __init__.py module into separate modules
