+++ doc/links.rst
@@ -21,7 +21,7 @@
 .. |License badge| image:: https://img.shields.io/badge/License-GPLv3-orange.svg
    :target: https://www.gnu.org/licenses/gpl-3.0
 
-.. |PyVersions badge| image:: https://img.shields.io/badge/python-3.11-blue.svg
+.. |PyVersions badge| image:: https://img.shields.io/badge/python-3.8_|_3.9_|_3.10_|_3.11|_3.12-blue.svg
    :target: https://github.com/btschwertfeger/python-cmethods
 
 .. |Downloads badge| image:: https://static.pepy.tech/personalized-badge/python-cmethods?period=total&units=abbreviation&left_color=grey&right_color=orange&left_text=downloads
