+++ doc/src/cmethods.rst
@@ -2,11 +2,13 @@
 Classes and Functions
 =====================
 
-.. currentmodule:: cmethods
+In past versions of the python-cmethods package (v1.x) there was a "CMethods"
+class that implemented the bias correction methods. This class was removed in
+version v2.0.0. Since then, the ``cmethods.adjust`` function is used to apply
+the implemented techniques except for detrended quantile mapping.
 
-.. autoclass:: CMethods
-   :members: adjust
-
+.. autofunction:: cmethods.adjust
+.. automethod:: cmethods.distribution.detrended_quantile_mapping
 
 Some additional methods
 -----------------------
