+++ doc/src/getting_started.rst
@@ -23,13 +23,13 @@ method specific documentation.
     :caption: Apply the Linear Scaling bias correction technique on 1-dimensional data
 
     import xarray as xr
-    from cmethods import CMethods as cm
+    from cmethods import adjust
 
     obsh = xr.open_dataset("input_data/observations.nc")
     simh = xr.open_dataset("input_data/control.nc")
     simp = xr.open_dataset("input_data/scenario.nc")
 
-    ls_result = cm.adjust(
+    ls_result = adjust(
         mathod="linear_scaling",
         obs=obsh["tas"][:, 0, 0],
         simh=simh["tas"][:, 0, 0],
@@ -42,13 +42,13 @@ method specific documentation.
     :caption: Apply the Quantile Delta Mapping bias correction technique on 3-dimensional data
 
     import xarray as xr
-    from cmethods import CMethods as cm
+    from cmethods import adjust
 
     obsh = xr.open_dataset("input_data/observations.nc")
     simh = xr.open_dataset("input_data/control.nc")
     simp = xr.open_dataset("input_data/scenario.nc")
 
-    qdm_result = cm.adjust(
+    qdm_result = adjust(
         method="quantile_delta_mapping",
         obs=obsh["tas"],
         simh=simh["tas"],
