+++ tests/test_methods.py
@@ -11,14 +11,11 @@ techniques.
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
-from cmethods.types import NPData_t, XRData_t
-
-if TYPE_CHECKING:
-    from cmethods import CMethods
-
 import pytest
+
+from cmethods import adjust
+from cmethods.distribution import detrended_quantile_mapping
+from cmethods.types import NPData_t, XRData_t
 
 from .helper import is_1d_rmse_better, is_3d_rmse_better
 
@@ -37,7 +34,6 @@ N_QUANTILES: int = 100
     ],
 )
 def test_1d_scaling(
-    cm: CMethods,
     datasets: dict,
     method: str,
     kind: str,
@@ -48,14 +44,12 @@ def test_1d_scaling(
     simp: XRData_t = datasets[kind]["simp"][:, 0, 0]
 
     # not group
-    result: XRData_t = cm.adjust(
-        method=method, obs=obsh, simh=simh, simp=simp, kind=kind
-    )
+    result: XRData_t = adjust(method=method, obs=obsh, simh=simh, simp=simp, kind=kind)
     assert isinstance(result, XRData_t)
     assert is_1d_rmse_better(result=result[kind], obsp=obsp, simp=simp)
 
     # grouped
-    result = cm.adjust(
+    result = adjust(
         method=method, obs=obsh, simh=simh, simp=simp, kind=kind, group=GROUP
     )
     assert isinstance(result, XRData_t)
@@ -73,7 +67,6 @@ def test_1d_scaling(
     ],
 )
 def test_3d_scaling(
-    cm: CMethods,
     datasets: dict,
     method: str,
     kind: str,
@@ -84,15 +77,19 @@ def test_3d_scaling(
     simp: XRData_t = datasets[kind]["simp"]
 
     # not grouped
-    result: XRData_t = cm.adjust(
-        method=method, obs=obsh, simh=simh, simp=simp, kind=kind
+    result: XRData_t = adjust(
+        method=method,
+        obs=obsh,
+        simh=simh,
+        simp=simp,
+        kind=kind,
     )
 
     assert isinstance(result, XRData_t)
     assert is_3d_rmse_better(result=result[kind], obsp=obsp, simp=simp)
 
     # grouped
-    result: XRData_t = cm.adjust(
+    result: XRData_t = adjust(
         method=method, obs=obsh, simh=simh, simp=simp, kind=kind, group=GROUP
     )
 
@@ -110,7 +107,6 @@ def test_3d_scaling(
     ],
 )
 def test_1d_distribution(
-    cm: CMethods,
     datasets: dict,
     method: str,
     kind: str,
@@ -120,7 +116,7 @@ def test_1d_distribution(
     simh: XRData_t = datasets[kind]["simh"][:, 0, 0]
     simp: XRData_t = datasets[kind]["simp"][:, 0, 0]
 
-    result: XRData_t = cm.adjust(
+    result: XRData_t = adjust(
         method=method,
         obs=obsh,
         simh=simh,
@@ -143,7 +139,6 @@ def test_1d_distribution(
     ],
 )
 def test_3d_distribution(
-    cm: CMethods,
     datasets: dict,
     method: str,
     kind: str,
@@ -153,7 +148,7 @@ def test_3d_distribution(
     simh: XRData_t = datasets[kind]["simh"]
     simp: XRData_t = datasets[kind]["simp"]
 
-    result: XRData_t = cm.adjust(
+    result: XRData_t = adjust(
         method=method,
         obs=obsh,
         simh=simh,
@@ -166,7 +161,7 @@ def test_3d_distribution(
     assert is_3d_rmse_better(result=result[kind], obsp=obsp, simp=simp)
 
 
-def test_detrended_quantile_mapping_add_1d(cm: CMethods, datasets: dict) -> None:
+def test_detrended_quantile_mapping_add_1d(datasets: dict) -> None:
     kind: str = "+"
     obsh: XRData_t = datasets[kind]["obsh"][:, 0, 0]
     obsp: XRData_t = datasets[kind]["obsp"][:, 0, 0]
@@ -174,14 +169,14 @@ def test_detrended_quantile_mapping_add_1d(cm: CMethods, datasets: dict) -> None
     simp: XRData_t = datasets[kind]["simp"][:, 0, 0]
 
     # not group
-    result: XRData_t = cm.detrended_quantile_mapping(
+    result: XRData_t = detrended_quantile_mapping(
         obs=obsh, simh=simh, simp=simp, kind=kind, n_quantiles=N_QUANTILES
     )
     assert isinstance(result, NPData_t)
     assert is_1d_rmse_better(result=result, obsp=obsp, simp=simp)
 
 
-def test_detrended_quantile_mapping_mult_1d(cm: CMethods, datasets: dict) -> None:
+def test_detrended_quantile_mapping_mult_1d(datasets: dict) -> None:
     kind: str = "*"
     obsh: XRData_t = datasets[kind]["obsh"][:, 0, 0]
     obsp: XRData_t = datasets[kind]["obsp"][:, 0, 0]
@@ -189,7 +184,7 @@ def test_detrended_quantile_mapping_mult_1d(cm: CMethods, datasets: dict) -> Non
     simp: XRData_t = datasets[kind]["simp"][:, 0, 0]
 
     # not group
-    result: XRData_t = cm.detrended_quantile_mapping(
+    result: XRData_t = detrended_quantile_mapping(
         obs=obsh, simh=simh, simp=simp, kind=kind, n_quantiles=N_QUANTILES
     )
     assert isinstance(result, NPData_t)
